+++ tket-exts/src/tket_exts/__init__.py
@@ -29,7 +29,7 @@ from tket_exts.tket.wasm import WasmExtension
 
 # This is updated by our release-please workflow, triggered by this
 # annotation: x-release-please-version
-__version__ = "0.14.1"
+__version__ = "0.14.2"
 
 __all__ = [
     "argument",